"2/3  les méthodes de Dijkstra (d'après de crous)"

+++ tp04/src/graph/MinDistance.java
@@ -1,7 +1,7 @@
 package graph;
 public interface MinDistance {
 	//changer la minDistance d'un sommet V par la distance d
-	public void replaceMinDistance(int d,Vertex V);
+	public void replaceMinDistance(float d,Vertex V);
 	
 	//retourner la minDistance d'un sommet V
 	public int actuelMinDistance(Vertex V);
